+++ lib/src/wire/resources/animation/animation.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/animation.dart' as animation;
+
+part 'curves.dart';