+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * text style bug fixed
 * text widget properties added
 * bottom app bar added
+* new builder for ChangeNotifier
+* IntrinsicHeight added
+* Curves added
 
 ## 0.0.2